fix #6 fix server crash with parseMetadata

+++ src/main/java/nf/fr/ephys/cookiecore/helpers/ChatHelper.java
@@ -55,7 +55,7 @@ public class ChatHelper {
 			return amount + "mB";
 		}
 
-		double converted = amount / 1000;
+		double converted = (double)amount / 1000;
 		if (converted < 1000) {
 			return MathHelper.round(converted, 3) + "B";
 		}
